Extracted _download_multimedia_paths_rows so it can be tested
